refactor: 패키지 레이어 네이밍 및 전역변수 wrapper 형식으로 변경
