+++ src/main/java/subway/application/PathService.java
@@ -3,8 +3,8 @@ package subway.application;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.dao.SectionDao;
-import subway.dao.StationDao;
+import subway.infrastrucure.SectionDao;
+import subway.infrastrucure.StationDao;
 import subway.domain.Fare;
 import subway.domain.ShortestPath;
 import subway.domain.Station;
