+++ src/main/java/subway/domain/Line.java
@@ -3,20 +3,18 @@ package subway.domain;
 import java.util.Objects;
 
 public class Line {
+
     private Long id;
-    private String name;
-    private String color;
+    private final String name;
+    private final String color;
 
-    public Line() {
-    }
-
-    public Line(String name, String color) {
+    public Line(Long id, String name, String color) {
+        this.id = id;
         this.name = name;
         this.color = color;
     }
 
-    public Line(Long id, String name, String color) {
-        this.id = id;
+    public Line(String name, String color) {
         this.name = name;
         this.color = color;
     }
