+++ src/main/java/subway/domain/Station.java
@@ -5,10 +5,7 @@ import java.util.Objects;
 public class Station {
 
     private Long id;
-    private String name;
-
-    public Station() {
-    }
+    private final String name;
 
     public Station(Long id, String name) {
         this.id = id;
