+++ src/main/java/subway/infrastrucure/SectionDao.java
@@ -1,4 +1,4 @@
-package subway.dao;
+package subway.infrastrucure;
 
 import java.util.List;
 import java.util.Map;
